Docs: split the playground component into a toolbar and hooks
